+++ examples/playground/src/env.d.ts
@@ -6,7 +6,7 @@ declare module '*.vue' {
   export default component;
 }
 
-declare module '@codemonster-ru/vueforge-codeblock' {
+declare module '@codemonster-ru/vueforge-codeblock/view' {
   import type { DefineComponent } from 'vue';
   export const VfCodeBlock: DefineComponent<Record<string, unknown>, Record<string, unknown>, unknown>;
 }